+++ extensions/git-repos/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Git Repos Changelog
 
+## [Enhancement] - 2022-04-20
+
+- Added a new action for the List Repos command to mark repos as favorites that will be displayed at the top.
+- Replaced custom caching logic with useCachedPromise.
+- Update to the latest Raycast API.
 ## [Enhancement] - 2022-11-25
 
 - Added a new action for List Repos command to open in all assigned applications at once.